Break out locustfile into several files
