+++ load_testing/apps/__init__.py
@@ -1,0 +1,4 @@
+from .anon import AnonBehavior  # noqa
+from .milmove import MilMoveUserBehavior  # noqa
+from .office import OfficeUserBehavior  # noqa
+from .tsp import TSPUserBehavior  # noqa
