+++ load_testing/apps/anon.py
@@ -1,0 +1,9 @@
+from locust import TaskSet
+from locust import task
+
+
+class AnonBehavior(TaskSet):
+
+    @task(1)
+    def index(self):
+        self.client.get("/")
