+++ load_testing/apps/office.py
@@ -1,0 +1,53 @@
+from locust import seq_task
+from bravado.client import SwaggerClient
+from bravado.requests_client import RequestsClient
+
+from .base import BaseTaskSequence
+from .base import InternalAPIMixin
+from .base import get_swagger_config
+
+
+class OfficeUserBehavior(BaseTaskSequence, InternalAPIMixin):
+
+    login_gov_user = None
+    session_token = None
+    user = {}
+
+    @seq_task(1)
+    def login(self):
+        resp = self.client.post('/devlocal-auth/create', data={"userType": "office"})
+        try:
+            self.login_gov_user = resp.json()
+            self.session_token = self.client.cookies.get('office_session_token')
+            self.requests_client = RequestsClient()
+            # Set the session to be the same session as locust uses
+            self.requests_client.session = self.client
+            # Set the csrf token in the global headers for all requests
+            # Don't validate requests or responses because we're using OpenAPI Spec 2.0 which doesn't respect
+            # nullable sub-definitions
+            self.swagger_internal = SwaggerClient.from_url(
+                "http://officelocal:8080/internal/swagger.yaml",
+                request_headers={'x-csrf-token': self.csrf},
+                http_client=self.requests_client,
+                config=get_swagger_config())
+        except Exception:
+            print(resp.content)
+
+    @seq_task(2)
+    def retrieve_user(self):
+        resp = self.client.get("/internal/users/logged_in")
+        self.user = resp.json()
+        # check response for 200
+
+    @seq_task(3)
+    def view_new_moves_queue(self):
+        self.swagger_internal_wrapper(
+            self.swagger_internal.queues.showQueue,
+            queueType="new")
+
+    @seq_task(4)
+    def logout(self):
+        self.client.post("/auth/logout")
+        self.login_gov_user = None
+        self.session_token = None
+        self.user = {}
